Final touches to make MSF happy with all refactorings
